% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/aggregateColumns.R
\name{aggregateColumns}
\alias{aggregateColumns}
\title{Aggregate Columns of a data frame}
\usage{
aggregateColumns(df, colums.of.interest, group.by)
}
\arguments{
\item{df}{source data frame}

\item{colums.of.interest}{list of columns to aggregate}

\item{group.by}{columns to group by}
}
\description{
Aggregate Columns of a data frame
}

