% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/aggregatedatasets.R
\name{aggregateDataset}
\alias{aggregateDataset}
\title{Aggregate Datasets}
\usage{
aggregateDataset(mean, stdev, rpl)
}
\arguments{
\item{mean}{List of mean values}

\item{stdev}{List of corresponding standard deviations}

\item{rpl}{List of corresponding replicate counts}
}
\description{
Till Dettmering, tilldettmering(at)gmail.com
}
\details{
based on http://books.google.de/books?id=cn3JtP5ugQsC&lpg=PA141&ots=YHV9PQJB9v&dq=varianz%20von%20%20zusammengefasste%20gruppen%20formel&hl=de&pg=PA141#v=onepage&q&f=false
}

