% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_helpers.R
\name{char.as.c}
\alias{char.as.c}
\title{Format the input vector as a character vector expression (c definition) for use in coding}
\usage{
char.as.c(x, unique.names = TRUE)
}
\description{
Format the input vector as a character vector expression (c definition) for use in coding
}
