% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_helpers.R
\name{colour.swatches}
\alias{colour.swatches}
\title{Plot colour swatches to the screen for a supplied list}
\usage{
colour.swatches(vector.of.colours)
}
\description{
Plot colour swatches to the screen for a supplied list
}
