% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_helpers.R
\name{darken.colour}
\alias{darken.colour}
\title{Darken a hex format RGB colour string by a specified percentage}
\usage{
darken.colour(hashhexstring, percentage = 0.5)

darken.colour(hashhexstring, percentage = 0.5)
}
\description{
Darken a hex format RGB colour string by a specified percentage

Lighten a hex format RGB colour string by a specified percentage
}
