% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dfBackup.R
\name{dfBackup}
\alias{dfBackup}
\title{Save csv backups of a list of dataframes}
\usage{
dfBackup(df, subdir = "")
}
\arguments{
\item{df}{list of strings representing data frames to be backed up}
}
\value{
2014-11-19_dataframeA.csv, 2014-11-19_dataframeB.csv
}
\description{
Save csv backups of a list of dataframes
}
\examples{
dfBackup(c('dataframeA', 'dataframeB'))
}

