% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/generateList.R
\name{generateList}
\alias{generateList}
\title{Generates list with all combinations of values in columns of a data frame}
\usage{
generateList(src, columns, rm.empty = TRUE, row.no = FALSE)
}
\arguments{
\item{src}{source data frame}

\item{columns}{list of columns of interest}
}
\description{
Generates list with all combinations of values in columns of a data frame
}

