% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_helpers.R
\name{grepp}
\alias{grepp}
\title{grep-like function that returns matches}
\usage{
grepp(
  pattern,
  x,
  ignore.case = FALSE,
  perl = FALSE,
  value = FALSE,
  fixed = FALSE,
  useBytes = FALSE,
  invert = FALSE
)
}
\description{
grep-like function that returns matches
}
