% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operations.R
\name{multiplyAcross}
\alias{multiplyAcross}
\title{Multiply a vector to each row or column of a matrix}
\usage{
multiplyAcross(mat, vec, across = "cols")
}
\arguments{
\item{mat}{matrix}

\item{vec}{vector to multiply matrix by}

\item{across}{one of c("cols", "rows")}
}
\value{
matrix with each row or column multiplied by input vector
}
\description{
Multiply a vector to each row or column of a matrix
}
