% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_helpers.R
\name{names.as.c}
\alias{names.as.c}
\title{Format the names of the object as a character vector expression (c definition) for use in coding}
\usage{
\method{names}{as.c}(x, unique.names = TRUE)
}
\description{
Format the names of the object as a character vector expression (c definition) for use in coding
}
