% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_operations.R
\name{normalizeMat}
\alias{normalizeMat}
\title{Normalize a matrix such that columns or rows sum to 1}
\usage{
normalizeMat(mat, sum = "cols")
}
\arguments{
\item{mat}{input matrix}

\item{sum}{one of c('cols','rows')}
}
\value{
normalized matrix
}
\description{
Normalize a matrix such that columns or rows sum to 1
}
