% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_helpers.R
\name{proportion.true}
\alias{proportion.true}
\title{shortcut to calculate proportion of elements that are TRUE with options for NA handling (also works for 0/1 coding)}
\usage{
proportion.true(x, na.handling = "keep")
}
\description{
shortcut to calculate proportion of elements that are TRUE with options for NA handling (also works for 0/1 coding)
}
