% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_helpers.R
\name{pv}
\alias{pv}
\title{return start of a vector or top left corner of a data-frame/matrix}
\usage{
pv(df, rows = 10, cols = NULL, tail = NULL, tail.rows = NULL, tail.cols = NULL)
}
\description{
return start of a vector or top left corner of a data-frame/matrix
}
