% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_helpers.R
\name{read.default}
\alias{read.default}
\title{read in data using my default settings}
\usage{
read.default(
  file,
  path = NULL,
  header = TRUE,
  sep = "\\t",
  quote = "",
  na.strings = "no na strings",
  comment.char = "",
  stringsAsFactors = FALSE,
  as.char = FALSE,
  ...
)
}
\description{
read in data using my default settings
}
