% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_png.R
\name{save_png}
\alias{save_png}
\title{Save base R plots to png}
\usage{
save_png(name, img_dir = NULL)
}
\arguments{
\item{name}{"Name of the plot}

\item{img_dir}{"Output directory of where to store the outputs"}
}
\value{
*.png
}
\description{
Save base R plots to png
}
