% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_operations.R
\name{str_split_vec}
\alias{str_split_vec}
\title{Split each string and return the ith substring for each item in a vector}
\usage{
str_split_vec(vec, pattern, i)
}
\arguments{
\item{vec}{a vector of strings}

\item{pattern}{what to split on}

\item{i}{index of substring to return for each item in vec}
}
\value{
a vector of strings
}
\description{
Split each string and return the ith substring for each item in a vector
}
