% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/theme_bw2.R
\name{theme_bw2}
\alias{theme_bw2}
\title{Theme for ggplot 2, modified from theme_bw}
\usage{
theme_bw2(base_size = 12, base_family = "")
}
\description{
Theme for ggplot 2, modified from theme_bw
}

