% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/ttest.R
\name{welch}
\alias{welch}
\title{Welch test with given mean, standard deviations and sample sizes
Till Dettmering, tilldettmering(at)gmail.com}
\usage{
welch(ctrl.x, ctrl.sd, ctrl.n, smpl.x, smpl.sd, smpl.n)
}
\arguments{
\item{ctrl.x}{Mean of control}

\item{ctrl.sd}{SD of control}

\item{ctrl.n}{Sample size of control}

\item{smpl.x}{Mean of sample}

\item{smpl.sd}{SD of sample}

\item{smpl.n}{Sample size of sample}

\item{rpl}{List of corresponding replicate counts}
}
\description{
Welch test with given mean, standard deviations and sample sizes
Till Dettmering, tilldettmering(at)gmail.com
}

