% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_helpers.R
\name{write.default}
\alias{write.default}
\title{write out data using my default settings}
\usage{
write.default(
  x,
  file,
  need.row.names = FALSE,
  need.col.names = TRUE,
  quote = FALSE,
  sep = "\\t"
)
}
\description{
write out data using my default settings
}
