% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateList.R
\name{generateList}
\alias{generateList}
\title{Generates list with all combinations of values in columns of a data frame}
\usage{
generateList(src, columns, rm.empty = TRUE, row.no = FALSE)
}
\arguments{
\item{src}{source data frame}

\item{columns}{list of columns of interest}

\item{rm.empty}{remove empty combinations. If this is set to FALSE, all possible combinations of all columns of interest will be returned.}

\item{row.no}{generate an extra column with the row name}
}
\description{
Generates list with all combinations of values in columns of a data frame
}

