% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Generates multiple plots in single view}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\description{
Generates multiple plots in single view
}

